Move the bbc parser + smiley parser to its own class, with stub functions left in place for now.
